# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "Swagger MyProfile",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "demo.intellifi.tech/demo/MyProfile/web/public",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/user/coordinate/setCoordinates": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Koordinat ekle",
        "operationId": "setCoordinates",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/UserSetCoordinates"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserCoordinates"
            }
          }
        }
      }
    },
    "/user/coordinate/nearbyUsers": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Yakındaki kullanıcılar",
        "operationId": "nearbyUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/NearbyUsers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/user/attendedActivity": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Etkinliğe katıldım",
        "operationId": "userAttendedActivity",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/UserAttendedActivities"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/user/userPrivacySettings": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Kullanıcının katıldığı etkinlikler",
        "operationId": "userPrivacySettings",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/UserPrivacySettings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserPrivacySettings"
            }
          }
        }
      }
    },
...
    "/user/stopFollowing": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Takibi bırak",
        "operationId": "stopFollowing",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/Follow"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "UserCoordinates": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "NearbyUsers": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        },
        "meterLimit": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserSetCoordinates": {
      "type": "object",
      "properties": {
        "api_token": {
          "type": "string"
        },
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserAttendedActivities": {
      "type": "object",
      "properties": {
        "api_token": {
          "type": "string"
        },
        "activity_id": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Follow": {
      "type": "object",
      "properties": {
        "from_user_id": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserPrivacySettings": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer"
        },
        "visibility_on_the_map": {
          "type": "boolean"
        },
        "no_message": {
          "type": "boolean"
        },
        "no_follow_up_request": {
          "type": "boolean"
        }
      },
      "xml": {
        "name": "User"
      }
    }
  },
...
}